mv index2.js -> index.js
